Improvements on version proof
